Gut the releases feature
We never used the functionality to add a release, mainly because it was 
never properly built.

This removes a ton of fields from the release model, and removes the 
Task model and association entirely.

The idea is that we can replace this soon with something that works 
better for us :)